Move GetWorkloadmetaInit to separate package
otel-agent and dogstatsd import cmd/agent/common only for this function
so move it to a separate package to reduce dependencies.
